Extend maximum expiration date
patch by Berenguer Blasi; reviewed by Andrés de la Peña for CASSANDRA-14227

+++ src/java/org/apache/cassandra/cql3/BatchQueryOptions.java
@@ -84,7 +84,7 @@ public abstract class BatchQueryOptions
         return wrapped.getTimestamp(state);
     }
 
-    public int getNowInSeconds(QueryState state)
+    public long getNowInSeconds(QueryState state)
     {
         return wrapped.getNowInSeconds(state);
     }
