+++ src/java/org/apache/cassandra/db/rows/UnfilteredSource.java
@@ -65,5 +65,5 @@ public interface UnfilteredSource
     long getMinTimestamp();
 
     /** Minimum local deletion time in the memtable */
-    int getMinLocalDeletionTime();
+    long getMinLocalDeletionTime();
 }
